Support mosn as a sidcar.

+++ bin/init.sh
@@ -90,19 +90,36 @@ function download_envoy_if_necessary () {
     mkdir -p "$(dirname "$2")"
     pushd "$(dirname "$2")"
 
-    # Download and extract the binary to the output directory.
-    echo "Downloading Envoy: ${DOWNLOAD_COMMAND} $1 to $2"
-    time ${DOWNLOAD_COMMAND} --header "${AUTH_HEADER:-}" "$1" | tar xz
+    if [ $SIDCAR = "Envoy" ] ; then     
+      # Download and extract the binary to the output directory.
+      echo "Downloading Envoy: ${DOWNLOAD_COMMAND} $1 to $2"
+      time ${DOWNLOAD_COMMAND} --header "${AUTH_HEADER:-}" "$1" | tar xz
 
-    # Copy the extracted binary to the output location
-    cp usr/local/bin/envoy "$2"
+      # Copy the extracted binary to the output location
+      cp usr/local/bin/envoy "$2"
 
-    # Remove the extracted binary.
-    rm -rf usr
+      # Remove the extracted binary.
+      rm -rf usr
 
-    # Make a copy named just "envoy" in the same directory (overwrite if necessary).
-    echo "Copying $2 to $(dirname "$2")/envoy"
-    cp -f "$2" "$(dirname "$2")/envoy"
+      # Make a copy named just "envoy" in the same directory (overwrite if necessary).
+      echo "Copying $2 to $(dirname "$2")/envoy"
+      cp -f "$2" "$(dirname "$2")/envoy"
+    fi  
+
+    if [ $SIDCAR = "MOSN" ] ; then     
+      # Download and extract the binary to the output directory.
+      echo "Downloading MOSN: ${DOWNLOAD_COMMAND} $1 to $2"
+      time ${DOWNLOAD_COMMAND} --header "${AUTH_HEADER:-}" "$1" > mosn
+
+      # Copy the extracted binary to the output location
+      cp mosn "$2"
+
+      # Remove the extracted binary.
+      rm -rf mosn
+
+      # Make a copy named just "mosn" in the same directory (overwrite if necessary).
+      echo "Copying $2 to $(dirname "$2")/mosn"
+    fi
     popd
   fi
 }
@@ -117,62 +134,101 @@ if [[ -z "${PROXY_REPO_SHA:-}" ]] ; then
   export PROXY_REPO_SHA
 fi
 
-# Defines the base URL to download envoy from
-ISTIO_ENVOY_BASE_URL=${ISTIO_ENVOY_BASE_URL:-https://storage.googleapis.com/istio-build/proxy}
-
-# These variables are normally set by the Makefile.
-# OS-neutral vars. These currently only work for linux.
-ISTIO_ENVOY_VERSION=${ISTIO_ENVOY_VERSION:-${PROXY_REPO_SHA}}
-ISTIO_ENVOY_DEBUG_URL=${ISTIO_ENVOY_DEBUG_URL:-${ISTIO_ENVOY_BASE_URL}/envoy-debug-${ISTIO_ENVOY_LINUX_VERSION}.tar.gz}
-ISTIO_ENVOY_RELEASE_URL=${ISTIO_ENVOY_RELEASE_URL:-${ISTIO_ENVOY_BASE_URL}/envoy-alpha-${ISTIO_ENVOY_LINUX_VERSION}.tar.gz}
-
-# Envoy Linux vars. Normally set by the Makefile.
-ISTIO_ENVOY_LINUX_VERSION=${ISTIO_ENVOY_LINUX_VERSION:-${ISTIO_ENVOY_VERSION}}
-ISTIO_ENVOY_LINUX_DEBUG_URL=${ISTIO_ENVOY_LINUX_DEBUG_URL:-${ISTIO_ENVOY_DEBUG_URL}}
-ISTIO_ENVOY_LINUX_RELEASE_URL=${ISTIO_ENVOY_LINUX_RELEASE_URL:-${ISTIO_ENVOY_RELEASE_URL}}
-# Variables for the extracted debug/release Envoy artifacts.
-ISTIO_ENVOY_LINUX_DEBUG_DIR=${ISTIO_ENVOY_LINUX_DEBUG_DIR:-"${OUT_DIR}/linux_amd64/debug"}
-ISTIO_ENVOY_LINUX_DEBUG_NAME=${ISTIO_ENVOY_LINUX_DEBUG_NAME:-"envoy-debug-${ISTIO_ENVOY_LINUX_VERSION}"}
-ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LINUX_DEBUG_PATH:-"${ISTIO_ENVOY_LINUX_DEBUG_DIR}/${ISTIO_ENVOY_LINUX_DEBUG_NAME}"}
-ISTIO_ENVOY_LINUX_RELEASE_DIR=${ISTIO_ENVOY_LINUX_RELEASE_DIR:-"${OUT_DIR}/linux_amd64/release"}
-ISTIO_ENVOY_LINUX_RELEASE_NAME=${ISTIO_ENVOY_LINUX_RELEASE_NAME:-"envoy-${ISTIO_ENVOY_LINUX_VERSION}"}
-ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LINUX_RELEASE_PATH:-"${ISTIO_ENVOY_LINUX_RELEASE_DIR}/${ISTIO_ENVOY_LINUX_RELEASE_NAME}"}
-
-# Envoy macOS vars. Normally set by the makefile.
-# TODO Change url when official envoy release for macOS is available
-ISTIO_ENVOY_MACOS_VERSION=${ISTIO_ENVOY_MACOS_VERSION:-1.0.2}
-ISTIO_ENVOY_MACOS_RELEASE_URL=${ISTIO_ENVOY_MACOS_RELEASE_URL:-https://github.com/istio/proxy/releases/download/${ISTIO_ENVOY_MACOS_VERSION}/istio-proxy-${ISTIO_ENVOY_MACOS_VERSION}-macos.tar.gz}
-# Variables for the extracted debug/release Envoy artifacts.
-ISTIO_ENVOY_MACOS_RELEASE_DIR=${ISTIO_ENVOY_MACOS_RELEASE_DIR:-"${OUT_DIR}/darwin_amd64/release"}
-ISTIO_ENVOY_MACOS_RELEASE_NAME=${ISTIO_ENVOY_MACOS_RELEASE_NAME:-"envoy-${ISTIO_ENVOY_MACOS_VERSION}"}
-ISTIO_ENVOY_MACOS_RELEASE_PATH=${ISTIO_ENVOY_MACOS_RELEASE_PATH:-"${ISTIO_ENVOY_MACOS_RELEASE_DIR}/${ISTIO_ENVOY_MACOS_RELEASE_NAME}"}
-
-# Allow override with a local build of Envoy
-USE_LOCAL_PROXY=${USE_LOCAL_PROXY:-0}
-if [[ ${USE_LOCAL_PROXY} == 1 ]] ; then
-  ISTIO_ENVOY_LOCAL_PATH=${ISTIO_ENVOY_LOCAL_PATH:-$(realpath "${ISTIO_GO}/../proxy/bazel-bin/src/envoy/envoy")}
-  echo "Using istio-proxy image from local workspace: ${ISTIO_ENVOY_LOCAL_PATH}"
-  if [[ ! -f "${ISTIO_ENVOY_LOCAL_PATH}" ]]; then
-    echo "Error: missing istio-proxy from local workspace: ${ISTIO_ENVOY_LOCAL_PATH}. Check your build path."
-    exit 1
-  fi
-
-  # Point the native paths to the local envoy build.
-  if [[ "$LOCAL_OS" == "Darwin" ]]; then
-    ISTIO_ENVOY_MACOS_RELEASE_PATH=${ISTIO_ENVOY_LOCAL_PATH}
-
-    ISTIO_ENVOY_LINUX_LOCAL_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH:-}
-    if [[ -f "${ISTIO_ENVOY_LINUX_LOCAL_PATH}" ]] ; then
-      ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH}
-      ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH}
-    else
-      echo "Warning: The specified local macOS Envoy will not be included by Docker images. Set ISTIO_ENVOY_LINUX_LOCAL_PATH to specify a custom Linux build."
+if [ $SIDCAR = "Envoy" ] ; then 
+  # Defines the base URL to download envoy from
+  ISTIO_ENVOY_BASE_URL=${ISTIO_ENVOY_BASE_URL:-https://storage.googleapis.com/istio-build/proxy}
+  
+  # These variables are normally set by the Makefile.
+  # OS-neutral vars. These currently only work for linux.
+  ISTIO_ENVOY_VERSION=${ISTIO_ENVOY_VERSION:-${PROXY_REPO_SHA}}
+  ISTIO_ENVOY_DEBUG_URL=${ISTIO_ENVOY_DEBUG_URL:-${ISTIO_ENVOY_BASE_URL}/envoy-debug-${ISTIO_ENVOY_LINUX_VERSION}.tar.gz}
+  ISTIO_ENVOY_RELEASE_URL=${ISTIO_ENVOY_RELEASE_URL:-${ISTIO_ENVOY_BASE_URL}/envoy-alpha-${ISTIO_ENVOY_LINUX_VERSION}.tar.gz}
+  
+  # Envoy Linux vars. Normally set by the Makefile.
+  ISTIO_ENVOY_LINUX_VERSION=${ISTIO_ENVOY_LINUX_VERSION:-${ISTIO_ENVOY_VERSION}}
+  ISTIO_ENVOY_LINUX_DEBUG_URL=${ISTIO_ENVOY_LINUX_DEBUG_URL:-${ISTIO_ENVOY_DEBUG_URL}}
+  ISTIO_ENVOY_LINUX_RELEASE_URL=${ISTIO_ENVOY_LINUX_RELEASE_URL:-${ISTIO_ENVOY_RELEASE_URL}}
+  # Variables for the extracted debug/release Envoy artifacts.
+  ISTIO_ENVOY_LINUX_DEBUG_DIR=${ISTIO_ENVOY_LINUX_DEBUG_DIR:-"${OUT_DIR}/linux_amd64/debug"}
+  ISTIO_ENVOY_LINUX_DEBUG_NAME=${ISTIO_ENVOY_LINUX_DEBUG_NAME:-"envoy-debug-${ISTIO_ENVOY_LINUX_VERSION}"}
+  ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LINUX_DEBUG_PATH:-"${ISTIO_ENVOY_LINUX_DEBUG_DIR}/${ISTIO_ENVOY_LINUX_DEBUG_NAME}"}
+  ISTIO_ENVOY_LINUX_RELEASE_DIR=${ISTIO_ENVOY_LINUX_RELEASE_DIR:-"${OUT_DIR}/linux_amd64/release"}
+  ISTIO_ENVOY_LINUX_RELEASE_NAME=${ISTIO_ENVOY_LINUX_RELEASE_NAME:-"envoy-${ISTIO_ENVOY_LINUX_VERSION}"}
+  ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LINUX_RELEASE_PATH:-"${ISTIO_ENVOY_LINUX_RELEASE_DIR}/${ISTIO_ENVOY_LINUX_RELEASE_NAME}"}
+  
+  # Envoy macOS vars. Normally set by the makefile.
+  # TODO Change url when official envoy release for macOS is available
+  ISTIO_ENVOY_MACOS_VERSION=${ISTIO_ENVOY_MACOS_VERSION:-1.0.2}
+  ISTIO_ENVOY_MACOS_RELEASE_URL=${ISTIO_ENVOY_MACOS_RELEASE_URL:-https://github.com/istio/proxy/releases/download/${ISTIO_ENVOY_MACOS_VERSION}/istio-proxy-${ISTIO_ENVOY_MACOS_VERSION}-macos.tar.gz}
+  # Variables for the extracted debug/release Envoy artifacts.
+  ISTIO_ENVOY_MACOS_RELEASE_DIR=${ISTIO_ENVOY_MACOS_RELEASE_DIR:-"${OUT_DIR}/darwin_amd64/release"}
+  ISTIO_ENVOY_MACOS_RELEASE_NAME=${ISTIO_ENVOY_MACOS_RELEASE_NAME:-"envoy-${ISTIO_ENVOY_MACOS_VERSION}"}
+  ISTIO_ENVOY_MACOS_RELEASE_PATH=${ISTIO_ENVOY_MACOS_RELEASE_PATH:-"${ISTIO_ENVOY_MACOS_RELEASE_DIR}/${ISTIO_ENVOY_MACOS_RELEASE_NAME}"}
+  
+  
+  # Allow override with a local build of Envoy
+  USE_LOCAL_PROXY=${USE_LOCAL_PROXY:-0}
+  if [[ ${USE_LOCAL_PROXY} == 1 ]] ; then
+    ISTIO_ENVOY_LOCAL_PATH=${ISTIO_ENVOY_LOCAL_PATH:-$(realpath "${ISTIO_GO}/../proxy/bazel-bin/src/envoy/envoy")}
+    echo "Using istio-proxy image from local workspace: ${ISTIO_ENVOY_LOCAL_PATH}"
+    if [[ ! -f "${ISTIO_ENVOY_LOCAL_PATH}" ]]; then
+      echo "Error: missing istio-proxy from local workspace: ${ISTIO_ENVOY_LOCAL_PATH}. Check your build path."
+      exit 1
     fi
-  else
-    ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LOCAL_PATH}
-    ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LOCAL_PATH}
+  
+    # Point the native paths to the local envoy build.
+    if [[ "$LOCAL_OS" == "Darwin" ]]; then
+      ISTIO_ENVOY_MACOS_RELEASE_PATH=${ISTIO_ENVOY_LOCAL_PATH}
+  
+      ISTIO_ENVOY_LINUX_LOCAL_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH:-}
+      if [[ -f "${ISTIO_ENVOY_LINUX_LOCAL_PATH}" ]] ; then
+        ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH}
+        ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LINUX_LOCAL_PATH}
+      else
+        echo "Warning: The specified local macOS Envoy will not be included by Docker images. Set ISTIO_ENVOY_LINUX_LOCAL_PATH to specify a custom Linux build."
+      fi
+    else
+      ISTIO_ENVOY_LINUX_DEBUG_PATH=${ISTIO_ENVOY_LOCAL_PATH}
+      ISTIO_ENVOY_LINUX_RELEASE_PATH=${ISTIO_ENVOY_LOCAL_PATH}
+    fi
   fi
+
 fi
+
+
+if [ $SIDCAR = "MOSN" ] ; then 
+  # Defines the base URL to download mosn from
+  ISTIO_MOSN_BASE_URL=${ISTIO_MOSN_BASE_URL:-https://github.com/mosn/mosn/releases/download/}
+  
+  # These variables are normally set by the Makefile.
+  # OS-neutral vars. These currently only work for linux.
+  ISTIO_MOSN_VERSION=${ISTIO_MOSN_VERSION:-${PROXY_REPO_SHA}}
+  ISTIO_MOSN_DEBUG_URL=${ISTIO_MOSN_DEBUG_URL:-${ISTIO_MOSN_BASE_URL}/${ISTIO_MOSN_VERSION}/mosn}
+  ISTIO_MOSN_RELEASE_URL=${ISTIO_MOSN_RELEASE_URL:-${ISTIO_MOSN_BASE_URL}/${ISTIO_MOSN_VERSION}/mosn}
+  
+  # MOSN Linux vars. Normally set by the Makefile.
+  ISTIO_MOSN_LINUX_VERSION=${ISTIO_MOSN_LINUX_VERSION:-${ISTIO_MOSN_VERSION}}
+  ISTIO_MOSN_LINUX_DEBUG_URL=${ISTIO_MOSN_LINUX_DEBUG_URL:-${ISTIO_MOSN_DEBUG_URL}}
+  ISTIO_MOSN_LINUX_RELEASE_URL=${ISTIO_MOSN_LINUX_RELEASE_URL:-${ISTIO_MOSN_RELEASE_URL}}
+  # Variables for the extracted debug/release MOSN artifacts.
+  ISTIO_MOSN_LINUX_DEBUG_DIR=${ISTIO_MOSN_LINUX_DEBUG_DIR:-"${OUT_DIR}/linux_amd64/debug"}
+  ISTIO_MOSN_LINUX_DEBUG_NAME=${ISTIO_MOSN_LINUX_DEBUG_NAME:-"mosn-debug-${ISTIO_MOSN_LINUX_VERSION}"}
+  ISTIO_MOSN_LINUX_DEBUG_PATH=${ISTIO_MOSN_LINUX_DEBUG_PATH:-"${ISTIO_MOSN_LINUX_DEBUG_DIR}/${ISTIO_MOSN_LINUX_DEBUG_NAME}"}
+  ISTIO_MOSN_LINUX_RELEASE_DIR=${ISTIO_MOSN_LINUX_RELEASE_DIR:-"${OUT_DIR}/linux_amd64/release"}
+  ISTIO_MOSN_LINUX_RELEASE_NAME=${ISTIO_MOSN_LINUX_RELEASE_NAME:-"mosn-${ISTIO_MOSN_LINUX_VERSION}"}
+  ISTIO_MOSN_LINUX_RELEASE_PATH=${ISTIO_MOSN_LINUX_RELEASE_PATH:-"${ISTIO_MOSN_LINUX_RELEASE_DIR}/${ISTIO_MOSN_LINUX_RELEASE_NAME}"}
+  
+  # MOSN macOS vars. Normally set by the makefile.
+  # TODO Change url when official MOSN release for macOS is available
+  ISTIO_MOSN_MACOS_VERSION=${ISTIO_MOSN_MACOS_VERSION:-"${ISTIO_MOSN_VERSION}"}
+  ISTIO_MOSN_MACOS_RELEASE_URL=${ISTIO_MOSN_MACOS_RELEASE_URL:-https://github.com/mosn/mosn/releases/download/${ISTIO_MOSN_MACOS_VERSION}/mosn}
+  # Variables for the extracted debug/release MOSN artifacts.
+  ISTIO_MOSN_MACOS_RELEASE_DIR=${ISTIO_MOSN_MACOS_RELEASE_DIR:-"${OUT_DIR}//release"}
+  ISTIO_MOSN_MACOS_RELEASE_NAME=${ISTIO_MOSN_MACOS_RELEASE_NAME:-"mosn-${ISTIO_MOSN_MACOS_VERSION}"}
+  ISTIO_MOSN_MACOS_RELEASE_PATH=${ISTIO_MOSN_MACOS_RELEASE_PATH:-"${ISTIO_MOSN_MACOS_RELEASE_DIR}/${ISTIO_MOSN_MACOS_RELEASE_NAME}"}
+
+fi 
+
 
 mkdir -p "${ISTIO_OUT}"
 mkdir -p "${ISTIO_BIN}"
@@ -180,28 +236,43 @@ mkdir -p "${ISTIO_BIN}"
 # Set the value of DOWNLOAD_COMMAND (either curl or wget)
 set_download_command
 
+
 # Download and extract the Envoy linux debug binary.
-download_envoy_if_necessary "${ISTIO_ENVOY_LINUX_DEBUG_URL}" "$ISTIO_ENVOY_LINUX_DEBUG_PATH"
+#download_envoy_if_necessary "${ISTIO_SIDCAR_LINUX_DEBUG_URL}" "$ISTIO_SIDCAR_LINUX_DEBUG_PATH"
+download_envoy_if_necessary `eval echo '$ISTIO_'"${SIDCAR}"'_LINUX_DEBUG_URL'` `eval echo '$ISTIO_'"${SIDCAR}"'_LINUX_DEBUG_PATH'`
 
 # Download and extract the Envoy linux release binary.
-download_envoy_if_necessary "${ISTIO_ENVOY_LINUX_RELEASE_URL}" "$ISTIO_ENVOY_LINUX_RELEASE_PATH"
+download_envoy_if_necessary `eval echo '$ISTIO_'"${SIDCAR}"'_LINUX_RELEASE_URL'` `eval echo '$ISTIO_'"${SIDCAR}"'_LINUX_RELEASE_PATH'`
 
 if [[ "$LOCAL_OS" == "Darwin" ]]; then
   # Download and extract the Envoy macOS release binary
-  download_envoy_if_necessary "${ISTIO_ENVOY_MACOS_RELEASE_URL}" "$ISTIO_ENVOY_MACOS_RELEASE_PATH"
-  ISTIO_ENVOY_NATIVE_PATH=${ISTIO_ENVOY_MACOS_RELEASE_PATH}
+  download_envoy_if_necessary `eval echo '$ISTIO_'"${SIDCAR}"'_MACOS_RELEASE_URL'` `eval echo '$ISTIO_'"${SIDCAR}"'_MACOS_RELEASE_PATH'`
+  ISTIO_SIDCAR_NATIVE_PATH=`eval echo '$ISTIO_'"${SIDCAR}"'_MACOS_RELEASE_PATH'`
 else
-  ISTIO_ENVOY_NATIVE_PATH=${ISTIO_ENVOY_LINUX_DEBUG_PATH}
+  ISTIO_SIDCAR_NATIVE_PATH=`eval echo '$ISTIO_'"${SIDCAR}"'_LINUX_DEBUG_PATH'`
 fi
 
-# Copy native envoy binary to ISTIO_OUT
-echo "Copying ${ISTIO_ENVOY_NATIVE_PATH} to ${ISTIO_OUT}/envoy"
-cp -f "${ISTIO_ENVOY_NATIVE_PATH}" "${ISTIO_OUT}/envoy"
+if [ $SIDCAR = "Envoy" ] ; then 
+  # Copy native envoy binary to ISTIO_OUT
+  echo "Copying ${ISTIO_SIDCAR_NATIVE_PATH} to ${ISTIO_OUT}/envoy"
+  cp -f "${ISTIO_SIDCAR_NATIVE_PATH}" "${ISTIO_OUT}/envoy"
+  
+  # TODO(nmittler): Remove once tests no longer use the envoy binary directly.
+  # circleCI expects this in the bin directory
+  # Make sure the envoy binary exists. This is only used for tests, so use the debug binary.
+  echo "Copying ${ISTIO_OUT}/envoy to ${ISTIO_BIN}/envoy"
+  cp -f "${ISTIO_OUT}/envoy" "${ISTIO_BIN}/envoy"
+fi
 
-# TODO(nmittler): Remove once tests no longer use the envoy binary directly.
-# circleCI expects this in the bin directory
-# Make sure the envoy binary exists. This is only used for tests, so use the debug binary.
-echo "Copying ${ISTIO_OUT}/envoy to ${ISTIO_BIN}/envoy"
-cp -f "${ISTIO_OUT}/envoy" "${ISTIO_BIN}/envoy"
-
+if [ $SIDCAR = "MOSN" ] ; then 
+  # Copy native mosn binary to ISTIO_OUT
+  echo "Copying ${ISTIO_SIDCAR_NATIVE_PATH} to ${ISTIO_OUT}/mosn"
+  cp -f "${ISTIO_SIDCAR_NATIVE_PATH}" "${ISTIO_OUT}/mosn"
+  
+  # TODO(nmittler): Remove once tests no longer use the envoy binary directly.
+  # circleCI expects this in the bin directory
+  # Make sure the mosn binary exists. This is only used for tests, so use the debug binary.
+  echo "Copying ${ISTIO_OUT}/mosn to ${ISTIO_BIN}/mosn"
+  cp -f "${ISTIO_OUT}/mosn" "${ISTIO_BIN}/mosn"
+fi
 "${ROOTDIR}/bin/init_helm.sh"
